lang-portal/backend-flask/requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate>=4.0.5
invoke>=2.2.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-flask>=1.3.0
SQLAlchemy==2.0.23